+++ clients/derivatives_trading_usds_futures/src/binance_sdk_derivatives_trading_usds_futures/rest_api/models/modify_order_response.py
@@ -32,15 +32,9 @@ class ModifyOrderResponse(BaseModel):
     status: Optional[StrictStr] = None
     client_order_id: Optional[StrictStr] = Field(default=None, alias="clientOrderId")
     price: Optional[StrictStr] = None
-    avg_price: Optional[StrictStr] = Field(
-        default=None,
-        description="Average execution price. Will be removed after CM migration.",
-        alias="avgPrice",
-    )
     orig_qty: Optional[StrictStr] = Field(default=None, alias="origQty")
     executed_qty: Optional[StrictStr] = Field(default=None, alias="executedQty")
     cum_qty: Optional[StrictStr] = Field(default=None, alias="cumQty")
-    cum_base: Optional[StrictStr] = Field(default=None, alias="cumBase")
     time_in_force: Optional[StrictStr] = Field(default=None, alias="timeInForce")
     type: Optional[StrictStr] = None
     reduce_only: Optional[StrictBool] = Field(default=None, alias="reduceOnly")
@@ -73,11 +67,9 @@ class ModifyOrderResponse(BaseModel):
         "status",
         "clientOrderId",
         "price",
-        "avgPrice",
         "origQty",
         "executedQty",
         "cumQty",
-        "cumBase",
         "timeInForce",
         "type",
         "reduceOnly",
@@ -164,11 +156,9 @@ class ModifyOrderResponse(BaseModel):
                 "status": obj.get("status"),
                 "clientOrderId": obj.get("clientOrderId"),
                 "price": obj.get("price"),
-                "avgPrice": obj.get("avgPrice"),
                 "origQty": obj.get("origQty"),
                 "executedQty": obj.get("executedQty"),
                 "cumQty": obj.get("cumQty"),
-                "cumBase": obj.get("cumBase"),
                 "timeInForce": obj.get("timeInForce"),
                 "type": obj.get("type"),
                 "reduceOnly": obj.get("reduceOnly"),